requirements.txt
google-generativeai
fastapi
uvicorn
weaviate-client
python-dotenv
pydantic
faiss-cpu
PyPDF2
sentence-transformers
numpy
PyMuPDF
langchain-community